pyproject.toml
[project]
name = "rocky"
version = "0.1.0"
description = ""
authors = [
    {name = "Gopi (GKS)",email = "gopikrishnans1996@gmail.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "ollama (>=0.6.1,<0.7.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
extraPaths = []
